manageFeedRules/requirements.txt
requests>=2.31.0
click>=8.1.0
pyyaml>=6.0
tqdm>=4.65.0
pandas>=2.0.0
python-dateutil>=2.8.0